Moved some files and cleaned up banner designs
