+++ src/components/AssetAccountDetails/AssetAccountDetails.tsx
@@ -4,18 +4,18 @@ import type { AccountId, AssetId } from '@shapeshiftoss/caip'
 import { fromAssetId } from '@shapeshiftoss/caip'
 import { useMemo } from 'react'
 
-import { AccountAssets } from './AccountAssets/AccountAssets'
-import { AssetChart } from './AssetHeader/AssetChart'
-import { AssetDescription } from './AssetHeader/AssetDescription'
-import { AssetHeader } from './AssetHeader/AssetHeader'
-import { AssetMarketData } from './AssetHeader/AssetMarketData'
-import { Display } from './Display'
-import { Equity } from './Equity/Equity'
-import { Main } from './Layout/Main'
-import { MaybeChartUnavailable } from './MaybeChartUnavailable'
-import { RelatedAssets } from './RelatedAssets/RelatedAssets'
-import { SpamWarningBanner } from './SpamWarning/SpamWarningBanner'
-import { EarnOpportunities } from './StakingVaults/EarnOpportunities'
+import { AccountAssets } from '../AccountAssets/AccountAssets'
+import { AssetChart } from '../AssetHeader/AssetChart'
+import { AssetDescription } from '../AssetHeader/AssetDescription'
+import { AssetHeader } from '../AssetHeader/AssetHeader'
+import { AssetMarketData } from '../AssetHeader/AssetMarketData'
+import { Display } from '../Display'
+import { Equity } from '../Equity/Equity'
+import { Main } from '../Layout/Main'
+import { MaybeChartUnavailable } from '../MaybeChartUnavailable'
+import { RelatedAssets } from '../RelatedAssets/RelatedAssets'
+import { EarnOpportunities } from '../StakingVaults/EarnOpportunities'
+import { SpamWarningBanner } from './components/SpamWarningBanner'
 
 import { AssetTransactionHistory } from '@/components/TransactionHistory/AssetTransactionHistory'
 import { getChainAdapterManager } from '@/context/PluginProvider/chainAdapterSingleton'
